workers/shared-worker-parse-error-failure.html

<!DOCTYPE html>
<title>SharedWorker: parse error failure</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

const checkArguments = (args) => {
  assert_equals(args.a.constructor, Event);
  assert_true('message' in args.a, 'Event.message');
  assert_equals(typeof args.a.message, 'string', 'Event.message');
  assert_equals(args.a.filename, location.href + '/resources/syntax-error.js',
                'Event.filename');
  assert_equals(args.a.lineno, 1, 'Event.lineno');
  assert_equals(args.b, undefined, 'unexpected second argument to onerror');
  assert_equals(args.c, undefined, 'unexpected third argument to onerror');
};

promise_test(async () => {
  const scriptURL = 'modules/resources/syntax-error.js';
  const worker = new SharedWorker(scriptURL, { type: 'classic' });
  const args = await new Promise(resolve =>
      worker.onerror = (a, b, c) => resolve({a, b, c}));
  checkArguments(args);
}, 'Classic shared worker construction for script with syntax error should ' +
   'dispatch an event named error.');

promise_test(async () => {
  const scriptURL = 'modules/resources/static-import-worker.js';
  const worker = new SharedWorker(scriptURL, { type: 'classic' });
  const args = await new Promise(resolve =>
      worker.onerror = (a, b, c) => resolve({a, b, c}));
  checkArguments(args);
}, 'Static import on classic shared worker should dispatch an event named ' +
   'error.');

</script>
